requirements.txt
bcrypt
blinker
certifi
charset-normalizer
click
coverage
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-WTF
greenlet
idna
iniconfig
itsdangerous
Jinja2
MarkupSafe
packaging
pluggy
psycopg2-binary
PyJWT
pytest
python-dotenv
requests
SQLAlchemy
typing_extensions
urllib3
waitress
Werkzeug
schematics>=2.0.1
